fix logreturns data types

+++ contracts/core/IPriceConsumer.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.8.0;
 
 interface IPriceConsumer {
-  function getLatestPrice(address _feed) external view returns (int);
+  function getLatestPrice(address _feed) external view returns (int128);
 }
